+++ libs/sdk-core/CHANGELOG.md
@@ -9,6 +9,10 @@
 - Support input/output max tokens on `/ask` endpoint
 - Add support for `neighbouring_paragraphs` RAG strategy
 
+### Bug fix
+
+- Return `IErrorResponse` when ask request is successful but its response contains an error   
+
 # 1.20.0 (2024-09-02)
 
 ### Feature